Remove assertions on allocated and freed memory
The numbers are platform dependant because of the different width of
native strings and are hard to calculate manually.
They also change with every little change in the managed code so that
the tests become a burden.
Anyways don't remove the underlying feature as it might coe handy when
looking for palloc memory leaks in managed code.